confirmation, delete base service

+++ src/app/shared/component/crud-list.component.ts
@@ -2,6 +2,7 @@ import { Router } from '@angular/router';
 import { Injector, OnInit } from '@angular/core';
 import { MenuItem } from '../model/menu-item';
 import { BaseService } from '../service/base.service';
+import Swal from 'sweetalert2/dist/sweetalert2.js';
 
 export abstract class CrudListComponent<T> implements OnInit {
     list: T[];
@@ -11,7 +12,7 @@ export abstract class CrudListComponent<T> implements OnInit {
     constructor(protected service: BaseService<T>,
         protected injector: Injector,
         protected endpoint = '') {
-            this.router = this.injector.get(Router);
+        this.router = this.injector.get(Router);
     }
 
     ngOnInit(): void {
@@ -25,11 +26,26 @@ export abstract class CrudListComponent<T> implements OnInit {
     }
 
     remover(id: number): void {
-        this.service.deleteById(id).subscribe( () => {
-            // Swal.fire('Ok', 'Excluído com sucesso', 'success');
+        Swal.fire({
+            title: 'Você tem certeza?',
+            text: 'Essa operação é irreversível.',
+            icon: 'warning',
+            buttons: true,
+            dangerMode: true,
+        }).then(isConfirm => {
+            if (isConfirm) {
+                this.service.deleteById(id).subscribe(() => {
+                    Swal.fire('Ok', 'Excluído com sucesso', 'success');
+                });
+            }
         });
     }
 
+    /**
+     * Para indexar as listas, ganho de performance no ngFor
+     * @param index index do item do array
+     * @param item item do array
+     */
     trackById(index: number, item: T): void {
         return item['id'];
     }
